requirements.txt
aider_chat==0.48.1
GitPython==3.1.43
pandas==2.2.2
python-dotenv==1.0.1
python-socketio==5.11.3
requests==2.32.3
st_pages==1.0.1
streamlit==1.35.0
streamlit_option_menu==0.3.13
tokencost==0.1.11
